+++ Kokoro/docs/release-notes.md
@@ -6,6 +6,7 @@
 
 * For players and modders: Fixed `APlaySpecificCardFromAnywhere` (and in turn `APlayRandomCardsFromAnywhere`) getting into an endless loop if the hand ends up being full in between cards being played (apparent in the Soggins' Do Something! card).
 * For modders: Added `IStatusLogicHook.OnStatusTurnTrigger` and `IStatusLogicHook.HandleStatusTurnAutoStep` APIs.
+* For modders: Added resource cost action APIs.
 
 ## 1.1.1
 Released 28 December 2023.